--- libs/blocks/global-navigation/blocks/profile/button.css
@@ -1,11 +1,0 @@
-.feds-profile-img {
-  max-height: 30px;
-  max-width: 30px;
-  display: block;
-}
-
-.feds-profile-button {
-  padding: 5px var(--feds-gutter);
-  border-radius: var(--feds-radius-utilityIcon);
-  border: none;
-}
