--- libs/blocks/global-navigation/blocks/profile/signIn.css
@@ -1,46 +1,0 @@
-.feds-signIn {
-  padding: 11px var(--feds-gutter);
-  border-radius: var(--feds-radius-utilityIcon);
-  color: #4B4B4B;
-  white-space: nowrap;
-}
-
-.feds-signIn-dropdown {
-  position: absolute;
-  display: none;
-  right: 0;
-  top: 100%;
-  border: 1px solid #e1e1e1;
-  border-radius: 0 0 4px 4px;
-  background: #fafafa;
-  overflow: hidden;
-  box-shadow: 0 2px 6px -1px rgb(0 0 0 / 10%);
-  line-height: 1.4;
-  white-space: nowrap;
-  z-index: 1;
-}
-
-.feds-signIn[aria-expanded = "true"] + .feds-signIn-dropdown {
-  display: block;
-}
-
-[dir = "rtl"] .feds-signIn-dropdown {
-  right: initial;
-  left: 0;
-}
-
-.feds-signIn-dropdown ul {
-  margin: 0;
-  padding: 12px 0;
-}
-
-.feds-signIn-dropdown li > a {
-  display: block;
-  color: var(--feds-color-link--light);
-  padding: 6px 32px;
-}
-
-.feds-signIn-dropdown li > a:hover {
-  color: var(--feds-color-link--hover--light);
-  background-color: var(--feds-background-link--hover--light);;
-}
